refactor: Also move tests to world::reflect
